+++ app/store/configStore.js
@@ -1,0 +1,10 @@
+/*
+  Created by Uncle Charlie, 2016/11/15
+
+  @flow
+*/
+import {applyMiddleware, createStore} from 'redux';
+
+
+
+var isDebuggingInChrome = __DEV__ && !!window.navigator.userAgent
